+++ enums.py
@@ -42,16 +42,16 @@ class UpscaledCropPadStrategy(__BaseEnum):
 	how to tweak the image to get there:
 
 
-	• pad - upscale the image to fit it into the desired frame on one side,
+	• pad only - upscale the image to fit it into the desired frame on one side,
 	then add missing pixels on the other one for outpaint.
-	• crop - upscale the image to fill the entire frame, then crop extra pixels on one side.
+	• crop only - upscale the image to fill the entire frame, then crop extra pixels on one side.
 	• nearest - automatically choose one of the above, to crop/pad the least number of pixels (by area).
 	• exact-upscale - follow the provided upscale-value precisely. This is the only option
 	that uses it. Also, it's the only one that might require both outpainting and crop.
 
 	In case when both are required, the output values assume this order: upscale -> crop -> pad.
 	"""
-	PAD = 'pad'
-	CROP = 'crop'
+	PAD = 'pad only'
+	CROP = 'crop only'
 	NEAREST = 'nearest'
 	EXACT_UPSCALE = 'exact-upscale'