+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Prim-node rename (class and internal name, too): `Upscale Strategy (Best-Res)` -> `Crop-Pad Strategy (Best-Res)`
 - Inputs/outputs for upscale-related nodes are renamed to be more intuitive.
 - `Best-Res (area+scale)`: change (restore) the value returned by `upscale` output.
+- Upscale-related nodes: change some defaults. 
 
 # V1.1.0